kola/tests/kubernetes: reorganize cluster setup
Break out cluster setup from running the basic checks.
